java-toolchain: specify vendor and allow for automated resolution
This will spedify a specific version of jdk from a specified vendor and allow `gradle` to install a matching java version if not available on the host system.
