+++ settings.gradle.kts
@@ -1,3 +1,6 @@
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version("0.7.0")
+}
 rootProject.name = "beam-applications-java"
 
 include("etl")
